myitems component updated

+++ src/Pages/MyItems/MyItems.js
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
+import { toast, ToastContainer } from "react-toastify";
 import auth from "../../firebase.init";
+import 'react-toastify/dist/ReactToastify.css';
 
 const MyItems = () => {
   const [products, setProducts] = useState([]);
@@ -20,6 +22,10 @@ const MyItems = () => {
  
 
   const handleDelete = (id) => {
+    const confirm = window.confirm('Are you want to delete?');
+    if(!confirm){
+      return ;
+    }
     const url = `http://localhost:5000/fruits/${id}`;
     fetch(url, {
       method: "DELETE",
@@ -30,6 +36,7 @@ const MyItems = () => {
         if (data.deletedCount > 0) {
           const remaining = products.filter((product) => product._id != id);
           setProducts(remaining);
+          toast('Successfully deleted');
         }
       });
   };
@@ -67,6 +74,7 @@ const MyItems = () => {
           ))}
         </tbody>
       </table>
+      <ToastContainer />
     </div>
   );
 };
